Add base converter class to reduce code duplication
